Bootstrap -added bootstrap feature. -fix route query bug
