+++ package.json
@@ -9,9 +9,11 @@
   },
   "dependencies": {
     "axios": "^0.21.0",
+    "bootstrap": "^4.5.3",
     "cross-fetch": "^3.0.6",
     "next": "10.0.1",
     "react": "17.0.1",
+    "react-bootstrap": "^1.4.0",
     "react-dom": "17.0.1",
     "swr": "^0.3.8"
   }
