+++ pages/_app.js
@@ -1,4 +1,5 @@
-import '../styles/globals.css'
+// import '../styles/globals.css'
+import 'bootstrap/dist/css/bootstrap.min.css';
 
 function MyApp({ Component, pageProps }) {
   return <Component {...pageProps} />
