+++ pages/api/links.js
@@ -2,7 +2,7 @@ export default (req, res) => {
     res.statusCode = 200
     res.json([
         { href:"/useState", label:"useState Code" },
-        { href:"/routeLinks", label:"Route Links" },
+        { href:"/navItems", label:"Route Links" },
     ])
 }
