+++ pages/useState.js
@@ -1,4 +1,7 @@
-import { useState } from "react"
+import { useState } from 'react'
+import { Button } from 'react-bootstrap'
+import MainPage from '../components/MainPage'
+import styles from '../styles/Home.module.css'
 
 function UseState() {
     const [names, setName] = useState([]);
@@ -24,29 +27,39 @@ function UseState() {
     }
 
     return (
-       <div>
-           <main>
-                <h1>How to use useState()</h1>
-                <button 
-                // onClick={constFunc}
-                onClick={callFunc}
-                >
-                    Click me
-                </button>
-                
-                <ul>
+        <MainPage>
+            <h1>How to use useState()</h1>
+            
+            <div>
+                <div>
+                    <span className={styles.description}>View the code to learn</span>
+                </div>
+                <div>
+                    {/* <button 
+                        // onClick={constFunc}
+                        onClick={callFunc}
+                    >
+                        Click me
+                    </button> */}
+                    <Button
+                        variant='primary'
+                        onClick={callFunc}
+                        >
+                        Click Add
+                    </Button>
+                    <ul>
+                        {/* {names.map(({ id, name }) => { //explicit return
+                            <li key={id}>{name}</li>
+                        })} */}
 
-                    {/* {names.map(({ id, name }) => { //explicit return
-                        <li key={id}>{name}</li>
-                    })} */}
-
-                    {names.map(({ id, name }) => ( //implicit return
-                        <li key={id}>{name}</li>
-                    ))}
-
-                </ul>
-           </main>
-       </div>
+                        {names.map(({ id, name }) => ( //implicit return
+                            <li key={id}>{name}</li>
+                        ))}
+                    </ul>
+                </div>
+            </div>
+            
+        </MainPage>
     )
 }
